[W5.3b] 2. Extracted more variables and methods into Contact.java
